Improved imgui support
commit 26dbfee3f131359fb364c6ca4763be49ceefe0cc
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 13 17:34:14 2022 +0900

    wip

commit 1308f4b037b701c7effb1d7553c5c93a200289a3
Author: lriki <lriki.net@gmail.com>
Date:   Tue Jul 12 19:53:02 2022 +0900

    wip

commit 3e44af9a4f0da56ab7422c78ad337508aeb832ca
Author: lriki <lriki.net@gmail.com>
Date:   Tue Jul 12 14:38:58 2022 +0900

    wip

commit 3e7b6e6a441e1904bffa398c4148b5ad9122f008
Author: lriki <lriki.net@gmail.com>
Date:   Mon Jul 11 20:01:09 2022 +0900

    wip

commit 2befea61c3f004a41a2e9f46e279c7af6fa7fa89
Author: lriki <lriki.net@gmail.com>
Date:   Fri Jul 8 18:34:58 2022 +0900

    wip

commit 6a0ec08a88a4b9bbbed26d1a82e0d2780252d749
Author: lriki <lriki.net@gmail.com>
Date:   Thu Jul 7 17:49:19 2022 +0900

    wip

commit a6d812fb0f395330ff1a023b186883a803d45580
Author: lriki <lriki.net@gmail.com>
Date:   Thu Jul 7 13:17:23 2022 +0900

    wip

commit 71cc926ef71c1ddaa71d164a5cd21478e3995ba0
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 6 20:06:39 2022 +0900

    wip

commit ea20a5c07223025bdf2132e523604eeb1380c869
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 6 20:06:28 2022 +0900

    wip

commit 35f403d7ea44e5cfd9dfc931d167aaf6668aee2a
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 6 16:03:54 2022 +0900

    wip

commit e8b2aa1ff8c4d5098172756855003e80a387943a
Author: lriki <lriki.net@gmail.com>
Date:   Wed Jul 6 14:04:44 2022 +0900

    wip

commit 9b70073523164fc95c83aaad525c3a7dc5c9f628
Author: lriki <lriki.net@gmail.com>
Date:   Mon Jul 4 13:06:46 2022 +0900

    wip

+++ lumino/LuminoCore/src/Text/Encoding.cpp
@@ -284,7 +284,7 @@ String TextEncoding::decode(const byte_t* bytes, int length, int* outUsedDefault
     // 出力バッファのサイズを、実際に使用したバイト数にする
     output.resize(result.outputByteCount);
 
-    return String((Char*)output.data(), output.size() / sizeof(Char));
+    return String((Char*)output.data(), static_cast<int>(output.size() / sizeof(Char)));
 }
 
 std::vector<byte_t> TextEncoding::encode(const String& str, int* outUsedDefaultCharCount)
